Move components and fix build errors
