+++ site/src/modules/workspaces/WorkspaceMoreActions/useWorkspaceDuplication.ts
@@ -4,7 +4,7 @@ import { linkToTemplate, useLinks } from "modules/navigation";
 import { useCallback } from "react";
 import { useQuery } from "react-query";
 import { useNavigate } from "react-router-dom";
-import type { CreateWorkspaceMode } from "./CreateWorkspacePage";
+import type { CreateWorkspaceMode } from "../../../pages/CreateWorkspacePage/CreateWorkspacePage";
 
 function getDuplicationUrlParams(
 	workspaceParams: readonly WorkspaceBuildParameter[],
